refactor: modularize offline sync hooks into hooks/offline/
Move useOfflineSyncInit, useSyncStatus, useIsOnline, and useQueueAction
out of the flat useOfflineSync.ts into their own files under
hooks/offline/, with a barrel index.ts re-exporting all four. Update
every consumer (App.tsx, hooks/index.ts, OfflineIndicator.tsx,
useGroups.ts) to import from the new module. Add an integration test
covering the queue-then-flush cycle across useQueueAction and
useSyncStatus.

Closes #1261
