+++ frontend/src/hooks/offline/index.ts
@@ -1,0 +1,8 @@
+/**
+ * hooks/offline — barrel export for offline-first PWA hooks
+ */
+
+export { useOfflineSyncInit } from './useOfflineSyncInit';
+export { useSyncStatus } from './useSyncStatus';
+export { useIsOnline } from './useIsOnline';
+export { useQueueAction } from './useQueueAction';
